pyproject.toml
[tool.poetry]
name = "folding"
version = "0.1.0"
description = ""
authors = ["Andrew Herbig <notandrewherbig@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.11"

[tool.poetry.dev-dependencies]
pre-commit = "^2.9.3"
isort = "^5.6.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
